+++ components/keyscreen/emulate/src/main/java/com/flipperdevices/keyscreen/emulate/api/KeyEmulateApiImpl.kt
@@ -6,8 +6,8 @@ import com.flipperdevices.bridge.dao.api.model.FlipperKey
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.keyscreen.api.KeyEmulateApi
-import com.flipperdevices.keyscreen.emulate.composable.ComposableEmulate
-import com.flipperdevices.keyscreen.emulate.composable.ComposableSend
+import com.flipperdevices.keyscreen.emulate.composable.ComposableSimpleEmulateButton
+import com.flipperdevices.keyscreen.emulate.composable.ComposableSubGhzSendButton
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 
@@ -19,13 +19,13 @@ class KeyEmulateApiImpl @Inject constructor() : KeyEmulateApi {
 
         if (fileType != null) {
             when (fileType) {
-                FlipperKeyType.SUB_GHZ -> ComposableSend(
+                FlipperKeyType.SUB_GHZ -> ComposableSubGhzSendButton(
                     modifier = modifier,
                     flipperKey = flipperKey
                 )
                 FlipperKeyType.I_BUTTON,
                 FlipperKeyType.RFID,
-                FlipperKeyType.NFC -> ComposableEmulate(
+                FlipperKeyType.NFC -> ComposableSimpleEmulateButton(
                     modifier = modifier,
                     flipperKey = flipperKey
                 )