+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [Feature] Prepare for Fdroid app store
 - [Feature] Try to intent google assistant command with examples
 - [Feature] Handle errors while emulate/send
+- [Feature] New emulate/send buttons
 - [FIX] Now search for devices by Mac address and by flipper name
 - [FIX] Subghz provisioning while geoip is null